[newjersey/doula-pm#332] Remove legal feature flag
This commit removes the `NEXT_PUBLIC_FLAG_LEGAL` feature flag, and associated functionality and tests.

This commit resolves newjersey/doula-pm#332.
