+++ .env.test
@@ -1,3 +1,2 @@
 # Automatically loaded when running with NODE_ENV=test
 NEXT_PUBLIC_FLAG_TEST=1
-NEXT_PUBLIC_FLAG_LEGAL=1
