+++ cypress/.env.test-production-flags
@@ -1,2 +1,1 @@
 NEXT_PUBLIC_FLAG_TEST=0
-NEXT_PUBLIC_FLAG_LEGAL=0
