--- cypress/e2e/fillAndDownloadApplicationProductionFlags.cy.ts
@@ -1,9 +1,0 @@
-import {
-  baseExpectedFields,
-  baseFormPages,
-  fillAndDownloadApplication,
-} from "./fillAndDownloadApplication";
-
-it("should fill and download the application [productionFlags]", () => {
-  fillAndDownloadApplication(baseFormPages, baseExpectedFields);
-});
