+++ src/app/form/(formSteps)/legal/2/LegalStep2.test.tsx
@@ -21,15 +21,6 @@ import {
 import { screen } from "@testing-library/react";
 
 describe("<LegalStep2 />", () => {
-  const oldProcessEnv = process.env;
-  beforeEach(() => {
-    jest.resetModules();
-    process.env = { ...oldProcessEnv, NEXT_PUBLIC_FLAG_LEGAL: "1" };
-  });
-  afterAll(() => {
-    process.env = oldProcessEnv;
-  });
-
   const renderFunction = (dataStore: DataStore = {}) =>
     renderWithProviders(<LegalStep2 />, "/form/legal/2", dataStore);
