+++ src/app/form/_utils/formProgress.ts
@@ -47,17 +47,13 @@ export const getProgressBarSections = (): Array<Section> => {
       shouldShowProgressBar: true,
       shouldShowProgressHeadingAndRequiredMessage: true,
     },
-    ...(process.env.NEXT_PUBLIC_FLAG_LEGAL === "1"
-      ? [
-          {
-            id: "legal",
-            name: "Legal",
-            numSteps: 3,
-            shouldShowProgressBar: true,
-            shouldShowProgressHeadingAndRequiredMessage: true,
-          },
-        ]
-      : []),
+    {
+      id: "legal",
+      name: "Legal",
+      numSteps: 3,
+      shouldShowProgressBar: true,
+      shouldShowProgressHeadingAndRequiredMessage: true,
+    },
     {
       id: "review",
       name: "Review",
